# packages/order-service/src/routes/orders.ts
import { Router } from 'express';
import { z } from 'zod';
import mongoose from 'mongoose';
import { Order, IOrder } from '../models/Order';
import { logger } from '../config/logger';
import { checkIdempotency, setIdempotency } from '../services/idempotency';
import { validateCustomer, reserveStock } from '../services/validation';
import { generatePaymentToken } from '../services/paymentLink';
import type { OrderResponse } from '@microservice/shared';

export const ordersRouter = Router();

const createOrderSchema = z.object({
  customerId: z.string().refine((val) => mongoose.Types.ObjectId.isValid(val), {
    message: 'Invalid customer ID format',
  }),
  productId: z.string().refine((val) => mongoose.Types.ObjectId.isValid(val), {
    message: 'Invalid product ID format',
  }),
  quantity: z.number().int().positive(),
});

// Helper: Build order response with optional payment token
function buildOrderResponse(order: IOrder): OrderResponse {
  return {
    orderId: order._id.toString(),
    customerId: order.customerId,
    productId: order.productId,
    quantity: order.quantity,
    amount: order.amount,
    status: order.status,
    createdAt: order.createdAt.toISOString(),
    ...(order.status === 'pending' && order.paymentToken
      ? { paymentToken: order.paymentToken }
      : {}),
  };
}

// POST /orders — create order and return payment link
ordersRouter.post('/', async (req, res): Promise<void> => {
  try {
    const idempotencyKey = req.headers['idempotency-key'] as string;

    if (!idempotencyKey) {
      res.status(400).json({ error: 'idempotency-key header is required' });
      return;
    }

    // Check idempotency
    const existingOrderId = await checkIdempotency(idempotencyKey);
    if (existingOrderId) {
      const existingOrder = await Order.findById(existingOrderId);
      if (existingOrder) {
        logger.info(
          { orderId: existingOrderId, status: existingOrder.status },
          'Returning cached order (idempotent)'
        );
        const response = buildOrderResponse(existingOrder);
        res.json(response);
        return;
      }
    }

    const parsed = createOrderSchema.safeParse(req.body);
    if (!parsed.success) {
      res.status(400).json({ error: parsed.error.flatten() });
      return;
    }

    const { customerId, productId, quantity } = parsed.data;

    // Validate customer exists
    const customerValid = await validateCustomer(customerId);
    if (!customerValid) {
      res.status(404).json({ error: 'Customer not found' });
      return;
    }

    // Reserve stock and get product price
    const productData = await reserveStock(productId, quantity);
    if ('error' in productData) {
      res.status(productData.error === 'Product not found' ? 404 : 409).json({
        error: productData.error,
        ...(productData.available !== undefined && { available: productData.available }),
      });
      return;
    }

    // Calculate amount from product price
    const amount = productData.price * quantity;

    // Create order (status: pending, awaiting payment)
    const order = new Order({
      customerId,
      productId,
      quantity,
      amount,
      status: 'pending',
    });

    await order.save();

    // Generate payment token for new order
    const paymentToken = await generatePaymentToken(
      order._id.toString(),
      productId,
      quantity,
      amount
    );

    // Store payment token in order
    order.paymentToken = paymentToken;
    await order.save();

    // Store idempotency key
    await setIdempotency(idempotencyKey, order._id.toString());

    logger.info({ orderId: order._id, amount }, 'Order created');

    const response = buildOrderResponse(order);
    res.status(201).json(response);
  } catch (err) {
    logger.error({ err }, 'Failed to create order');
    res.status(500).json({ error: 'Internal server error' });
  }
});

// GET /orders — list orders with pagination and filters
ordersRouter.get('/', async (req, res): Promise<void> => {
  try {
    const {
      customerId,
      status,
      limit = '20',
      page = '1',
      sortBy = 'createdAt',
      sortOrder = 'desc',
    } = req.query;

    // Build filter
    const filter: Record<string, unknown> = {};
    if (customerId) {
      if (!mongoose.Types.ObjectId.isValid(customerId as string)) {
        res.status(400).json({ error: 'Invalid customerId format' });
        return;
      }
      filter.customerId = customerId;
    }
    if (status) {
      filter.status = status;
    }

    // Parse pagination
    const limitNum = Math.min(Math.max(parseInt(limit as string, 10), 1), 100);
    const pageNum = Math.max(parseInt(page as string, 10), 1);
    const skip = (pageNum - 1) * limitNum;

    // Build sort
    const sort: Record<string, 1 | -1> = {};
    sort[sortBy as string] = sortOrder === 'asc' ? 1 : -1;

    // Query with pagination
    const [orders, total] = await Promise.all([
      Order.find(filter).sort(sort).skip(skip).limit(limitNum),
      Order.countDocuments(filter),
    ]);

    const totalPages = Math.ceil(total / limitNum);

    const response = {
      data: orders.map(buildOrderResponse),
      pagination: {
        total,
        page: pageNum,
        limit: limitNum,
        totalPages,
        hasMore: pageNum < totalPages,
      },
    };

    res.json(response);
  } catch (err) {
    logger.error({ err }, 'Failed to list orders');
    res.status(500).json({ error: 'Internal server error' });
  }
});

// GET /orders/:id — get order details
ordersRouter.get('/:id', async (req, res): Promise<void> => {
  try {
    const orderId = req.params.id;

    // Validate ObjectId format
    if (!mongoose.Types.ObjectId.isValid(orderId)) {
      res.status(400).json({ error: 'Invalid order ID format' });
      return;
    }

    const order = await Order.findById(orderId);
    if (!order) {
      res.status(404).json({ error: 'Order not found' });
      return;
    }

    const response = buildOrderResponse(order);
    res.json(response);
  } catch (err) {
    logger.error({ err }, 'Failed to fetch order');
    res.status(500).json({ error: 'Internal server error' });
  }
});